+++ cameo-api-cpp/src/message/Message.h
@@ -302,7 +302,7 @@ namespace message {
 	namespace PortInfo {
 		constexpr const char* PORT = "port"; // int32
 		constexpr const char* STATUS = "status"; // string
-		constexpr const char* APPLICATION_NAME_ID = "applicationNameId"; // string
+		constexpr const char* APPLICATION = "application"; // string
 	}
 
 	namespace PortInfoListResponse {