+++ cameo-api/src/main/java/fr/ill/ics/cameo/Server.java
@@ -238,6 +238,10 @@ public class Server {
 		return impl.getApplicationInfos(name);		
 	}
 	
+	public List<Application.Port> getApplicationPorts() {
+		return impl.getApplicationPorts();
+	}
+	
 	/**
 	 * 
 	 * @param id